requirements.txt
absl-py==2.3.1
accelerate==0.20.3
addict==2.4.0
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
anyio==4.5.2
appdirs==1.4.4
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
astroid==3.2.4
asttokens==3.0.0
async-lru==2.0.4
async-timeout==5.0.1
attrs==25.3.0
babel==2.17.0
backcall==0.2.0
beautifulsoup4==4.13.4
black==21.4b2
bleach==6.1.0
cachetools==5.5.2
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
cloudpickle==3.1.1
comm==0.2.3
contourpy==1.1.1
cycler==0.12.1
datasets==3.1.0
debugpy==1.8.15
decorator==5.2.1
defusedxml==0.7.1
descartes==1.1.0
detectron2==0.6+cu111
diffusers==0.35.1
dill==0.3.8
docker-pycreds==0.4.0
einops==0.8.1
exceptiongroup==1.3.0
executing==2.2.0
fastjsonschema==2.21.1
filelock==3.16.1
fire==0.7.0
flake8==7.1.2
fonttools==4.57.0
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2024.9.0
ftfy==6.2.3
future==1.0.0
fvcore==0.1.5.post20221221
git-filter-repo==2.47.0
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.40.3
google-auth-oauthlib==1.0.0
grpcio==1.70.0
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.4
hydra-core==1.3.2
idna==3.10
imageio==2.35.1
importlib_metadata==8.5.0
importlib_resources==6.4.5
iniconfig==2.1.0
iopath==0.1.9
ipykernel==6.29.5
ipython==8.12.0
ipywidgets==8.1.7
isoduration==20.11.0
isort==5.13.2
jedi==0.19.2
Jinja2==3.1.6
joblib==1.4.2
json5==0.12.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.6
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.3.8
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
kiwisolver==1.4.7
llvmlite==0.31.0
lyft-dataset-sdk==0.0.8
Markdown==3.7
MarkupSafe==2.1.5
matplotlib==3.5.2
matplotlib-inline==0.1.7
mccabe==0.7.0
mistune==3.1.3
mmcv-full==1.4.0
mmdet==2.14.0
-e git+https://github.com/open-mmlab/mmdetection3d.git@f1107977dfd26155fc1f83779ee6535d2468f449#egg=mmdet3d
mmsegmentation==0.14.1
multidict==6.1.0
multiprocess==0.70.16
mypy_extensions==1.1.0
narwhals==1.42.1
nbclient==0.10.1
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==2.2
notebook==7.3.3
notebook_shim==0.2.4
numba==0.48.0
numpy==1.19.5
nuscenes-devkit==1.1.10
oauthlib==3.3.1
omegaconf==2.3.0
opencv-python==4.12.0.88
overrides==7.7.0
packaging==25.0
pandas==1.4.4
pandocfilters==1.5.1
parso==0.8.4
pathspec==0.12.1
pathtools==0.1.2
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
pkgutil_resolve_name==1.3.10
platformdirs==4.3.6
plotly==6.2.0
pluggy==1.5.0
plyfile==1.0.3
portalocker==3.0.0
prettytable==3.11.0
prometheus_client==0.21.1
prompt_toolkit==3.0.51
propcache==0.2.0
protobuf==4.25.8
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycocotools==2.0.7
pycodestyle==2.12.1
pycparser==2.22
pydot==4.0.1
pyflakes==3.2.0
Pygments==2.19.2
pylint==3.2.7
pyparsing==3.1.4
pyquaternion==0.9.9
pytest==8.3.5
python-dateutil==2.9.0.post0
python-json-logger==3.3.0
pytz==2025.2
PyWavelets==1.4.1
PyYAML==6.0.2
pyzmq==27.0.1
referencing==0.35.1
regex==2024.11.6
requests==2.32.4
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.20.1
rsa==4.9.1
safetensors==0.5.3
scikit-image==0.19.3
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.12.2
Send2Trash==1.8.3
sentry-sdk==2.34.1
setproctitle==1.3.6
Shapely==1.8.5
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
stack-data==0.6.3
tabulate==0.9.0
tensorboard==2.14.0
tensorboard-data-server==0.7.2
termcolor==2.4.0
terminado==0.18.1
terminaltables==3.1.10
threadpoolctl==3.5.0
tifffile==2023.7.10
timm==0.6.13
tinycss2==1.2.1
tokenizers==0.20.3
toml==0.10.2
tomli==2.2.1
tomlkit==0.13.3
torch==1.10.0+cu111
torchaudio==0.10.0+rocm4.1
torchvision==0.11.1+cu111
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.46.3
trimesh==2.35.39
types-python-dateutil==2.9.0.20241206
typing_extensions==4.5.0
uri-template==1.3.0
urllib3==2.2.3
wandb==0.15.8
wcwidth==0.2.13
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.6
widgetsnbextension==4.0.14
xxhash==3.5.0
yacs==0.1.8
yapf==0.40.1
yarl==1.15.2
zipp==3.20.2
